# Conflicts:
#	AVTUK_ETH_Version-f/coma.cpp
#	dialogs/journalsdialog.cpp
#	gen/eabstractprotocomchannel.cpp
#	gen/s2.cpp

+++ dialogs/cordialog.h
@@ -57,7 +57,7 @@ private:
 
 signals:
    void sendCom45(quint32*);
-   void sendCom50(quint16* adr, float *data);
+   void sendCom50(quint32* adr, float *data);
    void RS485WriteCorBd(information*, float*);
    void RS485ReadCorBd(information*);
    void CorReadRequest();
